Move jupiter_perps_steps to core
