+++ jupiter_modular/steps/auto_set_position_type.py
@@ -1,4 +1,4 @@
 """Select a LONG position type via JupiterPerpsFlow."""
 
 async def run(engine):
-    engine.jp.select_position_type("long")
+    await engine.jp.select_position_type("long")
